requirement.txt
pillow
pytesseract
numpy
glob2
regex
opencv-python
pandas